+++ Kuroba/app/src/main/java/com/github/adamantcheese/chan/core/site/sites/chan420/Chan420.java
@@ -65,7 +65,7 @@ public class Chan420
         }
 
         @Override
-        public String desktopUrl(Loadable loadable, int postNo) {
+        public String desktopUrl(Loadable loadable, long postNo) {
             if (loadable.isCatalogMode()) {
                 if (postNo > 0) {
                     return "https://boards.420chan.org/" + loadable.boardCode + "/thread/" + postNo;
