+++ Kuroba/app/src/main/java/com/github/adamantcheese/chan/ui/cell/CardPostCell.java
@@ -142,7 +142,7 @@ public class CardPostCell
             boolean inPopup,
             boolean highlighted,
             boolean selected,
-            int markedNo,
+            long markedNo,
             boolean showDivider,
             ChanSettings.PostViewMode postViewMode,
             boolean compact,
