+++ Kuroba/app/src/main/java/com/github/adamantcheese/chan/ui/cell/PostStubCell.java
@@ -153,7 +153,7 @@ public class PostStubCell
             boolean inPopup,
             boolean highlighted,
             boolean selected,
-            int markedNo,
+            long markedNo,
             boolean showDivider,
             ChanSettings.PostViewMode postViewMode,
             boolean compact,
